requirements.txt
# ---------------------------
# Core Backend
# ---------------------------
fastapi==0.110.0
uvicorn[standard]==0.27.1

# ---------------------------
# Inngest (background jobs)
# ---------------------------
inngest==0.3.7

# ---------------------------
# Environment & Utilities
# ---------------------------
python-dotenv==1.0.1
pydantic==2.6.1
typing-extensions>=4.9.0

# ---------------------------
# Vector Database
# ---------------------------
qdrant-client==1.8.0

# ---------------------------
# LLM Providers
# ---------------------------

google-generativeai==0.3.2

# ---------------------------
# RAG / Document Processing
# ---------------------------


# ---------------------------
# PDF Parsing
# ---------------------------
pypdf>=3.17.0

# ---------------------------
# HTTP & Async Support
# ---------------------------
httpx==0.26.0
anyio>=4.2.0

# ---------------------------
# Logging / Observability
# ---------------------------
rich>=13.7.0

typing-extensions